python-backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-multipart==0.0.22
python-jose[cryptography]==3.5.0
httpx==0.28.1
supabase==2.28.2
groq==0.37.1
langchain==1.2.12
langchain-groq==1.1.2
langchain-community==0.4.1
langchain-core==1.2.19
langgraph==1.1.2
langgraph-checkpoint-postgres==3.0.4
pydantic==2.12.5
python-dotenv==1.2.2
aiofiles==25.1.0
psycopg[binary,pool]==3.3.3
